+++ R/select.R
@@ -1,7 +1,11 @@
 #' Interactively Select Map Features
 #'
 #' @param x map to use
-#'
+#' @param platform one of \code{"leaflet"} or \code{"mapview"} to indicate
+#'          the type of map you would like to use for selection
+#' @param index \code{logical} with \code{index=TRUE} to indicate return
+#'          the index of selected features rather than the actual
+#'          selected features
 #' @param ... other arguments
 #'
 #' @example ./inst/examples/examples_select.R
@@ -11,21 +15,21 @@ selectFeatures = function(x, ...) {
 }
 
 #' @export
-selectFeatures.sf = function(x, platform = c("mapview", "leaflet"), ...) {
+selectFeatures.sf = function(
+  x,
+  platform = c("mapview", "leaflet"),
+  index = FALSE,
+  ...
+) {
 
   if (length(platform) > 1) platform = platform[1]
 
   x = mapview:::checkAdjustProjection(x)
-  x$edit_group = as.character(1:nrow(x))
-
-  addfun = switch(as.character(sf::st_dimension(sf::st_geometry(x)))[1],
-                  "0" = leaflet::addCircleMarkers,
-                  "1" = leaflet::addPolylines,
-                  "2" = leaflet::addPolygons)
+  x$edit_id = as.character(1:nrow(x))
 
   if (platform == "mapview") {
     m = mapview::mapview()@map
-    m = addfun(map = m, data = x, weight = 1, group = ~edit_group)
+    m = mapview::addFeatures(m, data=x, layerId=~x$edit_id)
     m = leaflet::fitBounds(m,
                            lng1 = as.numeric(sf::st_bbox(x)[1]),
                            lat1 = as.numeric(sf::st_bbox(x)[2]),
@@ -34,14 +38,21 @@ selectFeatures.sf = function(x, platform = c("mapview", "leaflet"), ...) {
     m = mapview::addHomeButton(map = m, ext = mapview:::createExtent(x))
   } else {
     m = leaflet::addTiles(leaflet::leaflet())
-    m = addfun(map = m, data = x, weight = 1, group = ~edit_group)
+    m = mapview::addFeatures(m, data=x, layerId=~x$edit_id)
   }
 
   ind = selectMap(m, ...)
 
-  indx = ind$group[as.logical(ind$selected)]
-  # todrop = "edit_group"
-  return(x[as.numeric(indx), !names(x) %in% "edit_group"])
+  indx = ind$id[as.logical(ind$selected)]
+  # todrop = "edit_id"
+
+  # when index argument is TRUE return index rather than actual features
+  if(index) {
+    return(as.numeric(indx))
+  }
+
+  # return selected features
+  return(x[as.numeric(indx), !names(x) %in% "edit_id"])
 }
 
 #' @export