+++ NEWS.md
@@ -1,3 +1,15 @@
+# mapedit 0.2.0
+
+* add Shiny module functionality
+* add selectFeatures function for easy selection of features from simple features (sf)
+* defaults to repeat mode in editMap()
+* removes circle Leaflet.draw tool by default in editMap()
+* use layerId instead of group for select
+* uses Viewer window for selectMap()
+* promote mapview to Imports
+* uses newly exported mapview::addFeatures()
+
+
 # mapedit 0.1.0
 
 **API breaking change**